:rocket: feat(jwt): jwt 시크릿 .env로 이동, postgresql 설정 #10 #11

+++ src/main/java/Backend/HIFI/user/follow/Follow.java
@@ -1,7 +1,6 @@
 package Backend.HIFI.user.follow;
 
 import Backend.HIFI.user.User;
-import lombok.Builder;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
@@ -13,7 +12,7 @@ import javax.persistence.*;
 public class Follow {
 
     @Id @GeneratedValue
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "FOLLOWER", nullable = false)
